Unnecessary inheritance removed, functions moved
