+++ dlt/destinations/impl/bigquery/bigquery.py
@@ -16,7 +16,6 @@ from dlt.common.destination.client import (
     RunnableLoadJob,
     SupportsStagingDestination,
     LoadJob,
-    WithTableReflection,
 )
 from dlt.common.json import json
 from dlt.common.runtime.signals import sleep
@@ -176,9 +175,7 @@ class BigQueryMergeJob(SqlMergeFollowupJob):
         return sql
 
 
-class BigQueryClient(
-    SqlJobClientWithStagingDataset, SupportsStagingDestination, WithTableReflection
-):
+class BigQueryClient(SqlJobClientWithStagingDataset, SupportsStagingDestination):
     def __init__(
         self,
         schema: Schema,
