+++ dlt/destinations/impl/databricks/databricks.py
@@ -14,7 +14,6 @@ from dlt.common.destination.client import (
     RunnableLoadJob,
     SupportsStagingDestination,
     LoadJob,
-    WithTableReflection,
 )
 from dlt.common.configuration.specs import (
     AwsCredentialsWithoutDefaults,
@@ -303,9 +302,7 @@ class DatabricksMergeJob(SqlMergeFollowupJob):
         """
 
 
-class DatabricksClient(
-    SqlJobClientWithStagingDataset, SupportsStagingDestination, WithTableReflection
-):
+class DatabricksClient(SqlJobClientWithStagingDataset, SupportsStagingDestination):
     def __init__(
         self,
         schema: Schema,
