+++ dlt/destinations/impl/dremio/dremio.py
@@ -9,7 +9,6 @@ from dlt.common.destination.client import (
     SupportsStagingDestination,
     FollowupJobRequest,
     LoadJob,
-    WithTableReflection,
 )
 from dlt.common.schema import TColumnSchema, Schema
 from dlt.common.schema.typing import TColumnType
@@ -96,7 +95,7 @@ class DremioLoadJob(RunnableLoadJob, HasFollowupJobs):
             """)
 
 
-class DremioClient(SqlJobClientWithStagingDataset, SupportsStagingDestination, WithTableReflection):
+class DremioClient(SqlJobClientWithStagingDataset, SupportsStagingDestination):
     def __init__(
         self,
         schema: Schema,
