+++ dlt/destinations/impl/mssql/mssql.py
@@ -3,7 +3,6 @@ from typing import Dict, Optional, Sequence, List, Any
 from dlt.common.destination.client import (
     FollowupJobRequest,
     PreparedTableSchema,
-    WithTableReflection,
 )
 from dlt.common.destination import DestinationCapabilitiesContext
 from dlt.common.schema import TColumnSchema, TColumnHint, Schema
@@ -77,7 +76,7 @@ class MsSqlMergeJob(SqlMergeFollowupJob):
         return SqlMergeFollowupJob._new_temp_table_name("#" + table_name, op, sql_client)
 
 
-class MsSqlJobClient(InsertValuesJobClient, WithTableReflection):
+class MsSqlJobClient(InsertValuesJobClient):
     def __init__(
         self,
         schema: Schema,
