+++ dlt/destinations/impl/redshift/redshift.py
@@ -19,7 +19,6 @@ from dlt.common.destination.client import (
     PreparedTableSchema,
     SupportsStagingDestination,
     LoadJob,
-    WithTableReflection,
 )
 from dlt.common.destination.capabilities import DestinationCapabilitiesContext
 from dlt.common.schema import TColumnSchema, TColumnHint, Schema
@@ -164,7 +163,7 @@ class RedshiftMergeJob(SqlMergeFollowupJob):
         )
 
 
-class RedshiftClient(InsertValuesJobClient, SupportsStagingDestination, WithTableReflection):
+class RedshiftClient(InsertValuesJobClient, SupportsStagingDestination):
     def __init__(
         self,
         schema: Schema,
