+++ dlt/destinations/impl/snowflake/snowflake.py
@@ -10,7 +10,6 @@ from dlt.common.destination.client import (
     RunnableLoadJob,
     CredentialsConfiguration,
     SupportsStagingDestination,
-    WithTableReflection,
 )
 from dlt.common.schema.utils import get_columns_names_with_prop
 from dlt.common.storages.file_storage import FileStorage
@@ -117,9 +116,7 @@ class SnowflakeLoadJob(RunnableLoadJob, HasFollowupJobs):
                 self._sql_client.execute_sql(f"REMOVE {stage_file_path}")
 
 
-class SnowflakeClient(
-    SqlJobClientWithStagingDataset, SupportsStagingDestination, WithTableReflection
-):
+class SnowflakeClient(SqlJobClientWithStagingDataset, SupportsStagingDestination):
     def __init__(
         self,
         schema: Schema,
